Move r.image-related classes to rece.image package
